requirements.txt
msal==1.31.0
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4
rich==13.7.1
